Move ament_python metadata extraction to augmentation module (#135)
This will give other extensions and users finer control over ROS package
identification, and aligns this extension better with the same process
as is performed in colcon-core and colcon-python-setup-py.

The only change in behavior should be that ROS packages which declare
their build_type as ament_python but which contain no setup.py files
will no longer be skipped with an error message. Instead, the package
will fail to build with a message stating that the setup.py could not be
found.